backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
python-dotenv==0.21.0
PyJWT==2.6.0
Werkzeug==3.0.1
boto3==1.34.0

